fix: AI持仓支持小数股，重置后与用户持仓完全一致
- AiTradeRecord.shares 改为 Float 支持小数股
- 初始化和重置时不再截断用户的小数股数

+++ backend/app/models/ai_trade_record.py
@@ -8,7 +8,7 @@ class AiTradeRecord(BaseModel):
 
     symbol = Column(String(20), nullable=False, index=True)
     action = Column(String(10), nullable=False)  # 'buy' | 'sell'
-    shares = Column(Integer, nullable=False)
+    shares = Column(Float, nullable=False)
     price = Column(Float, nullable=False)  # 执行时的股价
     trade_date = Column(Date, nullable=False)
     reason = Column(Text, nullable=True)
